+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 - [357](https://github.com/vegaprotocol/vegawallet/pull/357) - Warn without failing when no connection on version verification
 - [360](https://github.com/vegaprotocol/vegawallet/pull/360) - Attempt to enable ANSI colors in Windows terminal
 - [372](https://github.com/vegaprotocol/vegawallet/pull/372) - Remove wallet name serialisation to allow wallet file to be renamed
+- [370](https://github.com/vegaprotocol/vegawallet/pull/370) - Ensure unsupported output is correctly reported
 
 ## 0.9.2
 *2021-10-27*